make Nix env wrapper suggestion a total solution
The prior code is a partial solution.

The default standardized `nix-shell` behavior and expectation is that if `shell.nix` not found - `nix-shell` loads `default.nix` (which is always possible, since `default.nix` describes packade derivation, and package derivation always can be loaded as shell environemt by Nix), so there are a number of projects that do not have `shell.nix`, but, expect `default.nix` to be used instead. And shell.nix can be in ../ or further up the tree, that is solved by using a well-supported/used library function.